LayoutTests/storage/indexeddb/factory-deletedatabase-expected.txt

Test IndexedDB's webkitIndexedDB.deleteDatabase().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


webkitIndexedDB.open('database-to-delete')
db = event.target.result
db.setVersion('new version')
trans = event.target.result
PASS trans !== null is true
store = db.createObjectStore('storeName', null)
store.createIndex('indexName', '')
PASS store.indexNames.contains('indexName') is true
store.add('value', 'key')
db.transaction('storeName', webkitIDBTransaction.READ_WRITE)
store = transaction.objectStore('storeName')
store.get('key')
PASS event.target.result is "value"
db.setVersion('new version')
request = webkitIndexedDB.deleteDatabase('database-to-delete')
webkitIndexedDB.open('database-to-delete')
db = event.target.result
db.setVersion('new version')
PASS db.objectStoreNames.length is 0
PASS successfullyParsed is true

TEST COMPLETE
